+++ Core/ivy.xml
@@ -50,9 +50,6 @@
         <dependency conf="core->default" org="commons-validator" name="commons-validator" rev="1.6"/>
         <dependency conf="core->default" org="net.htmlparser.jericho" name="jericho-html" rev="3.3"/>
         <dependency org="com.squareup.okhttp" name="okhttp" rev="2.7.5"/>
-
-        <!-- https://mvnrepository.com/artifact/jfree/jfreechart -->
-        <dependency org="org.jfree" name="jfreechart" rev="1.0.19"/>
         
         <!-- map support for geolocation -->
         <dependency conf="core->default" org="org.jxmapviewer" name="jxmapviewer2" rev="2.4"/>
