+++ Core/nbproject/project.xml
@@ -480,14 +480,6 @@
                 <binary-origin>release/modules/ext/jxmapviewer2-2.4.jar</binary-origin>
             </class-path-extension>
             <class-path-extension>
-                <runtime-relative-path>ext/jfreechart-1.0.19.jar</runtime-relative-path>
-                <binary-origin>release/modules/ext/jfreechart-1.0.19.jar</binary-origin>
-            </class-path-extension>
-            <class-path-extension>
-                <runtime-relative-path>ext/jcommon-1.0.23.jar</runtime-relative-path>
-                <binary-origin>release/modules/ext/jcommon-1.0.23.jar</binary-origin>
-            </class-path-extension>
-            <class-path-extension>
                 <runtime-relative-path>ext/jdom-2.0.5-contrib.jar</runtime-relative-path>
                 <binary-origin>release/modules/ext/jdom-2.0.5-contrib.jar</binary-origin>
             </class-path-extension>
