+++ Core/src/org/sleuthkit/autopsy/casemodule/datasourcesummary/Bundle.properties
@@ -33,6 +33,7 @@ DataSourceSummaryDetailsPanel.acquisitionDetailsTextArea.text=
 DataSourceSummaryDetailsPanel.acquisitionDetailsLabel.text=Acquisition Details:
 DataSourceSummaryDetailsPanel.unallocatedSizeLabel.text=Unallocated Space:
 DataSourceSummaryDetailsPanel.unallocatedSizeValue.text=
+DataSourceSummaryCountsPanel.byMimeTypeLabel.text=Files by MIME Type
 DataSourceSummaryCountsPanel.byCategoryLabel.text=Files by Category
-DataSourceSummaryCountsPanel.resultsByTypeLabel.text=Results by Type
+DataSourceSummaryCountsPanel.jLabel1.text=Results by Type
 DataSourceSummaryUserActivityPanel.programsRunLabel.text=Top Programs Run
