+++ Core/src/org/sleuthkit/autopsy/casemodule/datasourcesummary/Bundle.properties-MERGED
@@ -1,7 +1,6 @@
 CTL_DataSourceSummaryAction=Data Source Summary
 DataSourceSummaryCountsPanel.ArtifactCountsTableModel.count.header=Count
 DataSourceSummaryCountsPanel.ArtifactCountsTableModel.type.header=Result Type
-DataSourceSummaryCountsPanel.byMimeTypeLabel.text=Files by MIME Type
 DataSourceSummaryCountsPanel.FilesByCategoryTableModel.all.row=All
 DataSourceSummaryCountsPanel.FilesByCategoryTableModel.allocated.row=Allocated
 DataSourceSummaryCountsPanel.FilesByCategoryTableModel.count.header=Count
@@ -16,8 +15,6 @@ DataSourceSummaryCountsPanel.FilesByMimeTypeTableModel.executables.row=Executabl
 DataSourceSummaryCountsPanel.FilesByMimeTypeTableModel.images.row=Images
 DataSourceSummaryCountsPanel.FilesByMimeTypeTableModel.type.header=File Type
 DataSourceSummaryCountsPanel.FilesByMimeTypeTableModel.videos.row=Videos
-DataSourceSummaryCountsPanel_FilesByMimeTypeTableModel_notAnalyzed_label=Not Analyzed
-DataSourceSummaryCountsPanel_FilesByMimeTypeTableModel_other_label=Other
 DataSourceSummaryDetailsPanel.getDataSources.error.text=Failed to get the list of datasources for the current case.
 DataSourceSummaryDetailsPanel.getDataSources.error.title=Load Failure
 DataSourceSummaryDetailsPanel.units.bytes=\ bytes
@@ -61,8 +58,9 @@ DataSourceSummaryDetailsPanel.acquisitionDetailsTextArea.text=
 DataSourceSummaryDetailsPanel.acquisitionDetailsLabel.text=Acquisition Details:
 DataSourceSummaryDetailsPanel.unallocatedSizeLabel.text=Unallocated Space:
 DataSourceSummaryDetailsPanel.unallocatedSizeValue.text=
+DataSourceSummaryCountsPanel.byMimeTypeLabel.text=Files by MIME Type
 DataSourceSummaryCountsPanel.byCategoryLabel.text=Files by Category
-DataSourceSummaryCountsPanel.resultsByTypeLabel.text=Results by Type
+DataSourceSummaryCountsPanel.jLabel1.text=Results by Type
 DataSourceSummaryDialog.window.title=Data Sources Summary
 DataSourceSummaryNode.column.dataSourceName.header=Data Source Name
 DataSourceSummaryNode.column.files.header=Files
