+++ Core/src/org/sleuthkit/autopsy/casemodule/datasourcesummary/DataSourceSummaryDetailsPanel.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.datasourcesummary.ui;
+package org.sleuthkit.autopsy.casemodule.datasourcesummary;
 
 import java.text.DecimalFormat;
 import java.util.logging.Level;
@@ -25,7 +25,6 @@ import javax.swing.table.DefaultTableModel;
 import org.apache.commons.lang3.StringUtils;
 import org.openide.util.NbBundle.Messages;
 import org.sleuthkit.autopsy.casemodule.Case;
-import org.sleuthkit.autopsy.datasourcesummary.datamodel.DataSourceDetailsSummary;
 import org.sleuthkit.datamodel.DataSource;
 import org.sleuthkit.datamodel.Image;
 import org.sleuthkit.datamodel.TskCoreException;
@@ -74,9 +73,9 @@ class DataSourceSummaryDetailsPanel extends javax.swing.JPanel {
             updateDetailsPanelData(null, null, null, null);
         } else {
             updateDetailsPanelData(dataSource,
-                    DataSourceDetailsSummary.getSizeOfUnallocatedFiles(dataSource),
-                    DataSourceDetailsSummary.getOperatingSystems(dataSource),
-                    DataSourceDetailsSummary.getDataSourceType(dataSource));
+                    DataSourceInfoUtilities.getSizeOfUnallocatedFiles(dataSource),
+                    DataSourceInfoUtilities.getOperatingSystems(dataSource),
+                    DataSourceInfoUtilities.getDataSourceType(dataSource));
         }
     }
