+++ Core/src/org/sleuthkit/autopsy/casemodule/datasourcesummary/NonEditableTableModel.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.datasourcesummary.ui;
+package org.sleuthkit.autopsy.casemodule.datasourcesummary;
 
 import javax.swing.table.DefaultTableModel;
