+++ Core/src/org/sleuthkit/autopsy/casemodule/datasourcesummary/ViewSummaryInformationAction.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.datasourcesummary.ui;
+package org.sleuthkit.autopsy.casemodule.datasourcesummary;
 
 import java.awt.Frame;
 import java.awt.event.ActionEvent;
@@ -24,7 +24,6 @@ import javax.swing.AbstractAction;
 import javax.swing.SwingUtilities;
 import org.openide.util.NbBundle.Messages;
 import org.openide.windows.WindowManager;
-import org.sleuthkit.autopsy.datasourcesummary.ui.Bundle;
 
 /**
  * ViewSummaryInformationAction action for opening a Data Sources Summary Dialog
