--- Core/src/org/sleuthkit/autopsy/datasourcesummary/datamodel/DataSourceDetailsSummary.java
@@ -1,178 +1,0 @@
-/*
- * Autopsy Forensic Browser
- *
- * Copyright 2020 Basis Technology Corp.
- * Contact: carrier <at> sleuthkit <dot> org
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.sleuthkit.autopsy.datasourcesummary.datamodel;
-
-import java.sql.SQLException;
-import java.util.logging.Level;
-import org.sleuthkit.autopsy.coreutils.Logger;
-import static org.sleuthkit.autopsy.datasourcesummary.datamodel.DataSourceInfoUtilities.getBaseQueryResult;
-import org.sleuthkit.datamodel.BlackboardArtifact;
-import org.sleuthkit.datamodel.BlackboardAttribute;
-import org.sleuthkit.datamodel.DataSource;
-import org.sleuthkit.datamodel.TskData;
-
-/**
- * Provides methods to query for data source overview details.
- */
-public class DataSourceDetailsSummary {
-
-    private static final Logger logger = Logger.getLogger(DataSourceDetailsSummary.class.getName());
-
-    /**
-     * Gets the size of unallocated files in a particular datasource.
-     *
-     * @param currentDataSource The data source.
-     *
-     * @return The size or null if the query could not be executed.
-     */
-    public static Long getSizeOfUnallocatedFiles(DataSource currentDataSource) {
-        if (currentDataSource == null) {
-            return null;
-        }
-
-        final String valueParam = "value";
-        final String countParam = "count";
-        String query = "SELECT SUM(size) AS " + valueParam + ", COUNT(*) AS " + countParam
-                + " FROM tsk_files"
-                + " WHERE " + DataSourceInfoUtilities.getMetaFlagsContainsStatement(TskData.TSK_FS_META_FLAG_ENUM.UNALLOC)
-                + " AND type<>" + TskData.TSK_DB_FILES_TYPE_ENUM.SLACK.getFileType()
-                + " AND type<>" + TskData.TSK_DB_FILES_TYPE_ENUM.VIRTUAL_DIR.getFileType()
-                + " AND dir_type<>" + TskData.TSK_FS_NAME_TYPE_ENUM.VIRT_DIR.getValue()
-                + " AND name<>''"
-                + " AND data_source_obj_id=" + currentDataSource.getId();
-
-        DataSourceInfoUtilities.ResultSetHandler<Long> handler = (resultSet) -> {
-            if (resultSet.next()) {
-                // ensure that there is an unallocated count result that is attached to this data source
-                long resultCount = resultSet.getLong(valueParam);
-                return (resultCount > 0) ? resultSet.getLong(valueParam) : null;
-            } else {
-                return null;
-            }
-        };
-        String errorMessage = "Unable to get size of unallocated files; returning null.";
-
-        return DataSourceInfoUtilities.getBaseQueryResult(query, handler, errorMessage);
-    }
-
-    /**
-     * Retrieves the concatenation of operating system attributes for a
-     * particular data source.
-     *
-     * @param dataSource The data source.
-     *
-     * @return The concatenated value or null if the query could not be
-     *         executed.
-     */
-    public static String getOperatingSystems(DataSource dataSource) {
-        if (dataSource == null) {
-            return null;
-        }
-
-        return getConcattedAttrValue(dataSource.getId(),
-                BlackboardArtifact.ARTIFACT_TYPE.TSK_OS_INFO.getTypeID(),
-                BlackboardAttribute.ATTRIBUTE_TYPE.TSK_PROG_NAME.getTypeID());
-    }
-
-    /**
-     * Retrieves the concatenation of data source usage for a particular data
-     * source.
-     *
-     * @param dataSource The data source.
-     *
-     * @return The concatenated value or null if the query could not be
-     *         executed.
-     */
-    public static String getDataSourceType(DataSource dataSource) {
-        if (dataSource == null) {
-            return null;
-        }
-
-        return getConcattedAttrValue(dataSource.getId(),
-                BlackboardArtifact.ARTIFACT_TYPE.TSK_DATA_SOURCE_USAGE.getTypeID(),
-                BlackboardAttribute.ATTRIBUTE_TYPE.TSK_DESCRIPTION.getTypeID());
-    }
-
-    /**
-     * Generates a string which is a concatenation of the value received from
-     * the result set.
-     *
-     * @param query              The query.
-     * @param valueParam         The parameter for the value in the result set.
-     * @param separator          The string separator used in concatenation.
-     * @param errorMessage       The error message if the result set could not
-     *                           be received.
-     * @param singleErrorMessage The error message if a single result could not
-     *                           be obtained.
-     *
-     * @return The concatenated string or null if the query could not be
-     *         executed.
-     */
-    private static String getConcattedStringsResult(String query, String valueParam, String separator, String errorMessage, String singleErrorMessage) {
-        DataSourceInfoUtilities.ResultSetHandler<String> handler = (resultSet) -> {
-            String toRet = "";
-            boolean first = true;
-            while (resultSet.next()) {
-                try {
-                    if (first) {
-                        first = false;
-                    } else {
-                        toRet += separator;
-                    }
-                    toRet += resultSet.getString(valueParam);
-                } catch (SQLException ex) {
-                    logger.log(Level.WARNING, singleErrorMessage, ex);
-                }
-            }
-
-            return toRet;
-        };
-
-        return getBaseQueryResult(query, handler, errorMessage);
-    }
-
-    /**
-     * Generates a concatenated string value based on the text value of a
-     * particular attribute in a particular artifact for a specific data source.
-     *
-     * @param dataSourceId    The data source id.
-     * @param artifactTypeId  The artifact type id.
-     * @param attributeTypeId The attribute type id.
-     *
-     * @return The concatenated value or null if the query could not be
-     *         executed.
-     */
-    private static String getConcattedAttrValue(long dataSourceId, int artifactTypeId, int attributeTypeId) {
-        final String valueParam = "concatted_attribute_value";
-        String query = "SELECT attr.value_text AS " + valueParam
-                + " FROM blackboard_artifacts bba "
-                + " INNER JOIN blackboard_attributes attr ON bba.artifact_id = attr.artifact_id "
-                + " WHERE bba.data_source_obj_id = " + dataSourceId
-                + " AND bba.artifact_type_id = " + artifactTypeId
-                + " AND attr.attribute_type_id = " + attributeTypeId;
-
-        String errorMessage = "Unable to execute query to retrieve concatted attribute values.";
-        String singleErrorMessage = "There was an error retrieving one of the results.  That result will be omitted from concatted value.";
-        String separator = ", ";
-        return getConcattedStringsResult(query, valueParam, separator, errorMessage, singleErrorMessage);
-    }
-
-    private DataSourceDetailsSummary() {
-    }
-}
